+++ src/main/java/com/posicube/assignment/querylog/application/service/RateLimiter.java
@@ -1,4 +1,4 @@
-package com.posicube.assignment.querylog.application;
+package com.posicube.assignment.querylog.application.service;
 
 public interface RateLimiter {
     /**
